Extract ChatGPT Codex tests
